refactor(tui): separate runtime message contracts
